requirements.txt
streamlit
transformers
torch
sounddevice
openai-whisper
scipy
sentencepiece